Move pull_request type definition to top-level Github module
Originally this type was only used in one module, but it's being used in more,
so I'm moving it to the top-level namespace so it's more centralized and can be
used in other modules more easily.

Plus, pull requests are a Github construct, not purely a labels construct, so it
makes more sense.