Adds check for existing labels before adding to avoid duplicates
This commit adds a check to the add labels function to compare the existing list
of labels to what is being added. If the label doesn't already exist, we add it;
otherwise the function is a no-op.

Although Github doesn't "reapply" the label, it does create another event on the
PR timeline, which is confusing. This should help de-clutter that.

Closes #5