Process payloads in a background task
This sets up backgrounding of tasks so we can process Github payloads outside of
the web request. This is done by leveraging Elixir's `Task` module, which spawns
a new process. This process is supervised by the app and retried if an error
occurs.

Inspiration was taken from the code-corps-api [repo]; indeed, most of the Task
implementation is a direct copy.

Closes #1

[repo]:
https://github.com/code-corps/code-corps-api/blob/065a300a613eb950dfb5b642b9dab7886a69d893/lib/code_corps/processor/processor.ex