Cache GitHub installation tokens in ETS using ConCache library
This sets up a token cache to store the installation tokens in Erlang's ETS.
This is an in-memory store, so it's very performant. We generate tokens using
GitHub's API. These are good for 60 minutes. Since we have to make an API call
to generate each one, it adds latency. Since they are good for more than one
use, it makes sense to cache these.

We use the [ConCache] library over bare ETS since it adds TTL features. This
automatically deletes tokens after 55 minutes, giving us more than enough window
to generate a new one.

We use the `get_or_store/3` function to get the value or store it in one go.
This makes sure our value is always up to date. When the TTL passes, ConCache
purges the value, and next time we need it `get_or_store/3` will make the API
request for a new value and set it before returning out of the function.

Closes #2

[ConCache]: https://github.com/sasa1977/con_cache