Fix double-label issue
The problem is that GitHub sends multiple PR events after a review is
submitted. For example, it'll get a review submitted event, and then
Custodian's own labeling will cause it to get another pr labeled event.
Because the matching was on the pr's state, it matched for that second
event, even when it shouldn't.

Here's an example timeline.

1. pull_request -> action = opened : adds `needs-review`
2. pull_request_review -> state = approved : adds `ready-to-merge`,
  removes others
3. pull_request -> action = labeled : adds `needs-review`

(3) happens because the state was still "open" which is what we were
matching on. The end result would be removing "needs-review", adding
"ready-to-merge" and then adding "needs-review".

The fix is to match on the action instead, which won't cause this double
action.

Closes #16