Update repo uninstall to look up by repo_id instead of installation_id
Several repos could be installed by the same installation_id. Removing a
single one in this case threw an error since `Bots.get_by/1` returned
more than one result.

The proper way to handle this is to iterate over the repositories that
need to be removed and do so by their repo_id instead.